Initial commit - implements only one test
