+++ test/e2e/Apps/BasicPowerShell/HelloCities_HttpStart/run.ps1
@@ -1,0 +1,9 @@
+using namespace System.Net
+
+param($Request, $TriggerMetadata)
+
+$InstanceId = Start-DurableOrchestration -FunctionName "HelloCitiesOrchestrator"
+Write-Host "Started orchestration with ID = '$InstanceId'"
+
+$Response = New-DurableOrchestrationCheckStatusResponse -Request $Request -InstanceId $InstanceId
+Push-OutputBinding -Name Response -Value $Response
