+++ test/e2e/Apps/BasicPowerShell/SayHello/function.json
@@ -1,0 +1,9 @@
+{
+  "bindings": [
+    {
+      "name": "name",
+      "type": "activityTrigger",
+      "direction": "in"
+    }
+  ]
+}
