+++ test/e2e/Apps/BasicPowerShell/SayHello/run.ps1
@@ -1,0 +1,3 @@
+param($name)
+
+"Hello $name!"
