+++ test/e2e/Apps/BasicPowerShell/nuget.config
@@ -1,0 +1,16 @@
+<configuration>
+  <packageSources>
+    <clear/>
+    <add key="local" value="./packages/" />
+    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
+  </packageSources>
+
+  <packageSourceMapping>
+    <packageSource key="nuget.org">
+      <package pattern="*" />
+    </packageSource>
+    <packageSource key="local">
+      <package pattern="Microsoft.Azure.WebJobs.Extensions.DurableTask" />
+    </packageSource>
+  </packageSourceMapping>
+</configuration>
