+++ test/e2e/Apps/BasicPowerShell/requirements.psd1
@@ -1,0 +1,9 @@
+# This file enables modules to be automatically managed by the Functions service.
+# See https://aka.ms/functionsmanageddependency for additional information.
+#
+@{
+    # For latest supported version, go to 'https://www.powershellgallery.com/packages/Az'. 
+    # To use the Az module in your function app, please uncomment the line below.
+    # 'Az' = '14.*'
+    'AzureFunctions.PowerShell.Durable.SDK' = '1.*'
+}
