+++ test/e2e/Tests/TestCollections/FunctionAppCollectionSequential.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+using Xunit;
+
+namespace Microsoft.Azure.Durable.Tests.DotnetIsolatedE2E;
+
+[CollectionDefinition(Constants.FunctionAppCollectionSequentialName, DisableParallelization = true)]
+public class FunctionAppCollectionSequential : ICollectionFixture<FunctionAppFixture>
+{
+    // This class has no code, and is never created. Its purpose is simply
+    // to be the place to apply [CollectionDefinition] and all the
+    // ICollectionFixture<> interfaces.
+}
