+++ test/e2e/Tests/Tests/DistributedTracingEntitiesTests.cs
@@ -37,6 +37,7 @@ public class DistributedTracingEntitiesTests
     [Fact]
     [Trait("MSSQL", "Skip")] // Entities are not implemented for the MSSQL backend
     [Trait("DTS", "Skip")] // Distributed tracing is currently not working in DTS
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task DistributedTracingEntitiesTest()
     {
         // Start an Activity here which will appear as the root of all other Activities generated by this test
@@ -77,6 +78,7 @@ public class DistributedTracingEntitiesTests
     [Fact]
     [Trait("MSSQL", "Skip")] // Entities are not implemented for the MSSQL backend
     [Trait("DTS", "Skip")] // Distributed tracing is currently not working in DTS
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task DistributedTracingEntitiesFromClientTest()
     {
         // Start an Activity here which will appear as the root of all other Activities generated by this test
