+++ test/e2e/Tests/Tests/DistributedTracingTests.cs
@@ -33,6 +33,7 @@ public class DistributedTracingTests
 
     [Fact]
     [Trait("DTS", "Skip")] // Distributed tracing is currently not working in DTS
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task DistributedTracingTest()
     {
         // Start Activity
