+++ test/e2e/Tests/Tests/ExternalEventTests.cs
@@ -24,6 +24,7 @@ public class ExternalEventTests
     // Test that sending an event to a running orchestrator waiting for an external event will complete successfully,
     // and sending an event to a completed instance will throw a FailedPrecondition RpcException with details error message.
     [Fact]
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task RaiseExternalEventTests()
     {
         using HttpResponseMessage response = await HttpHelpers.InvokeHttpTrigger("ExternalEventOrchestrator_HttpStart", "");
@@ -50,6 +51,7 @@ public class ExternalEventTests
 
     // Test that sending an event to a not-exist InstanceId will throw an NotFoundRpc Exception.
     [Fact]
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task NotFoundInstanceTest()
     {
         string jsonContent = JsonSerializer.Serialize("instance-does-not-exist-test");
