+++ test/e2e/Tests/Tests/LargeOutputOrchestratorTests.cs
@@ -22,6 +22,7 @@ public class LargeOutputOrchestratorTests
 
     [Theory]
     [InlineData(65)] // Provide a value slightly exceeding the 64 KB Azure Queue Storage limit to trigger use of blob storage instead at Azure Storage backend.
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task LargeOutputStatusQueryTests(int sizeInKB)
     {
         using HttpResponseMessage response = await HttpHelpers.InvokeHttpTriggerWithBody("LargeOutputOrchestrator_HttpStart", sizeInKB.ToString(), "application/json");
@@ -42,6 +43,7 @@ public class LargeOutputOrchestratorTests
     [Theory]
     [InlineData(4608)]// This value exceeds the default 4 MB, as the test sets the threshold to 6 MB.
     [Trait("DTS", "Skip")] 
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task DurableTaskClientWriteOutputTests(int sizeInKB)
     {
         using HttpResponseMessage response = await HttpHelpers.InvokeHttpTriggerWithBody("LargeOutputOrchestrator_HttpStart", sizeInKB.ToString(), "application/json");
