+++ test/e2e/Tests/Tests/PurgeInstancesTests.cs
@@ -18,6 +18,7 @@ public class PurgeInstancesTests
     }
 
     [Fact]
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task PurgeOrchestrationHistory_StartAndEnd_Succeeds()
     {
         DateTime purgeStartTime = DateTime.MinValue;
@@ -30,6 +31,7 @@ public class PurgeInstancesTests
     }
 
     [Fact]
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task PurgeOrchestrationHistory_Start_Succeeds()
     {
         DateTime purgeStartTime = DateTime.MinValue;
@@ -43,6 +45,7 @@ public class PurgeInstancesTests
 
     [Fact]
     [Trait("DTS", "Skip")] // Skip this test as there is a bug with current DTS backend, the createdTimeTo couldn't be null. 
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task PurgeOrchestrationHistory_End_Succeeds()
     {
         DateTime purgeStartTime = DateTime.MinValue;
@@ -56,6 +59,7 @@ public class PurgeInstancesTests
 
     [Fact]
     [Trait("DTS", "Skip")] // Skip this test as there is a bug with current DTS backend, the createdTimeTo couldn't be null. 
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task PurgeOrchestrationHistory_NoBoundaries_Succeeds()
     {
         DateTime purgeStartTime = DateTime.MinValue;
@@ -69,6 +73,7 @@ public class PurgeInstancesTests
 
     [Fact]
     [Trait("DTS", "Skip")] // Skip this test as there is a bug with current DTS backend, the createdTimeTo couldn't be null. 
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task PurgeOrchestrationHistoryAfterInvocation_Succeeds()
     {
         using HttpResponseMessage response = await HttpHelpers.InvokeHttpTrigger("HelloCities_HttpStart", "");
@@ -87,6 +92,7 @@ public class PurgeInstancesTests
 
     [Fact]
     [Trait("DTS", "Skip")] // Skip this test as there is a bug with current DTS backend, the createdTimeTo couldn't be null. 
+    [Trait("PowerShell", "Skip")] // Test not yet implemented in PowerShell
     public async Task PurgeAfterPurge_ZeroRows()
     {
         using HttpResponseMessage response = await HttpHelpers.InvokeHttpTrigger("HelloCities_HttpStart", "");
